+++ src/wui/interactive_base.cc
@@ -359,7 +359,7 @@ void InteractiveBase::draw_overlay(RenderTarget& dst) {
 			const std::string gametime(gametimestring(egbase().get_gametime(), true));
 			// NOCOM
 			const std::string gametime_text = as_condensed(gametime);
-			dst.blit(Vector2f(5, 5), UI::g_fh1->render(gametime_text)->texts[0]->image, BlendMode::UseAlpha,
+			dst.blit(Vector2i(5, 5), UI::g_fh1->render(gametime_text)->texts[0]->image, BlendMode::UseAlpha,
 			         UI::Align::kLeft);
 
 			static boost::format node_format("(%i, %i)");
@@ -372,7 +372,7 @@ void InteractiveBase::draw_overlay(RenderTarget& dst) {
 // NOCOM
 		const Image* rendered_text = UI::g_fh1->render(node_text)->texts[0]->image;
 
-		dst.blit(Vector2f(get_w() - 5, get_h() - rendered_text->height() - 5), rendered_text,
+		dst.blit(Vector2i(get_w() - 5, get_h() - rendered_text->height() - 5), rendered_text,
 		         BlendMode::UseAlpha, UI::Align::kRight);
 	}
 
@@ -382,7 +382,7 @@ void InteractiveBase::draw_overlay(RenderTarget& dst) {
 		// NOCOM
 		const Image* rendered_text = UI::g_fh1->render(as_condensed(
 			(fps_format % (1000.0 / frametime_) % (1000.0 / (avg_usframetime_ / 1000))).str()))->texts[0]->image;
-		dst.blit(Vector2f((get_w() - rendered_text->width()) / 2, 5), rendered_text);
+		dst.blit(Vector2i((get_w() - rendered_text->width()) / 2, 5), rendered_text);
 	}
 }
